add meal module

+++ src/restaurants/restaurants.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, Query, Req, UploadedFiles, UseGuards, UseInterceptors } from '@nestjs/common';
+import { Body, Controller, Delete, ForbiddenException, Get, Param, Post, Put, Query, Req, UploadedFiles, UseGuards, UseInterceptors } from '@nestjs/common';
 import { FilesInterceptor } from '@nestjs/platform-express';
 import { Query as ExpressQuery } from 'express-serve-static-core';
 import { CreateRestaurantDto } from './dto/create-restaurant.dto';
@@ -35,15 +35,18 @@ export class RestaurantsController {
     }
 
     @Put(':id')
+    @UseGuards(AuthGuard())
     async updateRestaurant(@Param('id') id: string, @Body() restaurant: UpdateRestaurantDto): Promise<Restaurant> {
         await this.restaurantsServices.findById(id)
         return this.restaurantsServices.updateById(id, restaurant)
     }
 
     @Delete(':id')
-    async deleteRestaurant(@Param('id') id: string): Promise<{ deleted: Boolean }> {
+    @UseGuards(AuthGuard())
+    async deleteRestaurant(@Param('id') id: string, @CurrentUser() user: User): Promise<{ deleted: Boolean }> {
         const restaurant = await this.restaurantsServices.findById(id)
-        const isDeleted = await this.restaurantsServices.deleteImages(restaurant.images)
+        if (restaurant.user.toString() !== user._id.toString()) throw new ForbiddenException('You can not delete this restaurant.')
+            const isDeleted = await this.restaurantsServices.deleteImages(restaurant.images)
         if (isDeleted) {
             await this.restaurantsServices.deleteById(id)
             return { deleted: true }
